refactor package uiutil and related classes
